Add files via upload
This is one of the first project that I have worked on using python. I tried to plot graph of per day covid cases in India and Pakistan. I have filtered out the data from data set where country was one of columns with other mostly being date for a span of 5 months and rows represented country and number of cases. I tried to filter row with India and Pakistan as country and then converted both the data structure to two columns one being date and other being number of cases. Since the cases were only increment of cases. I tried to get the number of cases per day.

With that I tried for loop, if else, structuring data frames, list, converting data types from str to number and date. I have added the code which also has other code that didnt work as it gives and idea about how we learn from trial and error. I would be happy to get feedback it would help me adjust my way of learning and thinking as well.